backend/realtime-scorer/requirements.txt
functions-framework
numpy
joblib
xgboost
pandas
scikit-learn
gunicorn
google-cloud-firestore